+++ src/app/Controllers/about.component.ts
@@ -1,7 +1,7 @@
 import { Component } from '@angular/core';
 import { CheckService } from '../Services/CheckService';
 import { Values } from '../values';
-import { ApiService } from '../Services/ApiService';
+import { ApiService } from '../Services/Api/ApiService';
 import { ElectronService } from 'ngx-electron';
 import { environment } from '../../environments/environment';
