+++ src/app/Services/MessageService.ts
@@ -5,7 +5,7 @@ import Swal from 'sweetalert2';
 import { NewMessageEvent } from '../Models/Events/NewMessageEvent';
 import { Conversation } from '../Models/Conversation';
 import { Message } from '../Models/Message';
-import { ConversationApiService } from './ConversationApiService';
+import { ConversationApiService } from './Api/ConversationApiService';
 import { map } from 'rxjs/operators';
 import { UploadService } from './UploadService';
 import { KahlaUser } from '../Models/KahlaUser';
@@ -27,11 +27,11 @@ import { NewMemberEvent } from '../Models/Events/NewMemberEvent';
 import { GroupConversation } from '../Models/GroupConversation';
 import { DissolveEvent } from '../Models/Events/DissolveEvent';
 import { HomeService } from './HomeService';
-import { GroupsApiService } from './GroupsApiService';
-import { FriendsApiService } from './FriendsApiService';
+import { GroupsApiService } from './Api/GroupsApiService';
+import { FriendsApiService } from './Api/FriendsApiService';
 import { ProbeService } from './ProbeService';
 import { ThemeService } from './ThemeService';
-import { FilesApiService } from './FilesApiService';
+import { FilesApiService } from './Api/FilesApiService';
 import { FileType } from '../Models/FileType';
 import { MessageFileRef } from '../Models/MessageFileRef';
