+++ src/app/app.module.ts
@@ -30,19 +30,19 @@ import { HomeComponent } from './Controllers/home.component';
 import { ShareComponent } from './Controllers/share.component';
 import { FileHistoryComponent } from './Controllers/file-history.component';
 // Services
-import { ApiService } from './Services/ApiService';
+import { ApiService } from './Services/Api/ApiService';
 import { ParamService } from './Services/ParamService';
 import { CacheService } from './Services/CacheService';
 import { CheckService } from './Services/CheckService';
 import { UploadService } from './Services/UploadService';
-import { AuthApiService } from './Services/AuthApiService';
-import { ConversationApiService } from './Services/ConversationApiService';
-import { FilesApiService } from './Services/FilesApiService';
-import { FriendsApiService } from './Services/FriendsApiService';
-import { GroupsApiService } from './Services/GroupsApiService';
+import { AuthApiService } from './Services/Api/AuthApiService';
+import { ConversationApiService } from './Services/Api/ConversationApiService';
+import { FilesApiService } from './Services/Api/FilesApiService';
+import { FriendsApiService } from './Services/Api/FriendsApiService';
+import { GroupsApiService } from './Services/Api/GroupsApiService';
 import { MessageService } from './Services/MessageService';
 import { InitService } from './Services/InitService';
-import { DevicesApiService } from './Services/DevicesApiService';
+import { DevicesApiService } from './Services/Api/DevicesApiService';
 import { ThemeService } from './Services/ThemeService';
 import { TimerService } from './Services/TimerService';
 import { HomeService } from './Services/HomeService';
